refactor: ActivityHome.kt -> ActivityMain.kt
